chore: Format using prettier
There's a mixture of tabs and spaces which messed with my indentation a bit so I added a prettier file.
It seems like most files were in tabs so I also went with tabs.

Feel free to discard this if you don't want it. It might cause a bunch of merge issues if there are a lot of PRs in progress so might be best to do this some other time.